Split view test into different user roles interacting with the system
